+++ src/test/java/com/epam/reportportal/testng/Constants.java
@@ -31,11 +31,13 @@ public class Constants {
 	static final String BASIC_URL = "basic_url";
 	static final String DEFAULT_UUID = "default_uuid";
 	static final String DEFAULT_NAME = "default_name";
+	static final String DEFAULT_NAME_DISPLAY_NAME_ANNOTATION = "default_display_name";
 	static final String DEFAULT_PROJECT = "default_project";
 	static final ItemAttributesRQ ATTRIBUTE = new ItemAttributesRQ("key", "value");
 	static final Set<ItemAttributesRQ> ATTRIBUTES = new HashSet<>(Collections.singleton(ATTRIBUTE));
 	static final Mode MODE = Mode.DEFAULT;
 	static final String DEFAULT_DESCRIPTION = "default_description";
+	static final String DEFAULT_ANNOTATION_DESCRIPTION = "default_annotation_description";
 	static final Long DEFAULT_TIME = 1515594836210L;
 
 }