+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- `@Description` annotation support, by @oleksandr-fomenko
+- `@DisplayName` annotation support, by @oleksandr-fomenko
+- Putting last error logs of tests to Items' description, by @utdacc
 ### Changed
 - Client version updated on [5.2.13](https://github.com/reportportal/client-java/releases/tag/5.2.13), by @HardNorth
 ### Removed